requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
crispy-bootstrap5==2024.2
cron-descriptor==1.4.5
cryptography==43.0.1
css-inline==0.14.1
defusedxml==0.8.0rc2
Django==5.1.1
django-allauth==65.0.2
django-autocomplete-light==3.11.0
django-celery-beat==2.7.0
django-crispy-forms==2.3
django-debug-toolbar==5.0.0a0
django-extensions==3.2.3
django-formtools==2.5.1
django-taggit==6.1.0
django-timezone-field==7.0
environs==11.0.0
factory_boy==3.3.1
Faker==30.1.0
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
markdown2==2.5.0
marshmallow==3.22.0
oauthlib==3.2.2
packaging==24.1
parsy==2.1
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
PyJWT==2.6.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
